% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_project.R
\name{add_qfeatures}
\alias{add_qfeatures}
\title{Add qfeatures to dataList}
\usage{
add_qfeatures(
  dataList,
  qfeatures,
  fileAnnotation = NULL,
  siriusFileColumnName = "ClassyFire superclass"
)
}
\arguments{
\item{dataList}{Output from readProjectData.}

\item{qfeatures}{qfeature object, can be taken from resultObj$qfeatures.}

\item{fileAnnotation}{character Path for sirius annotation file.}

\item{siriusFileColumnName}{One of "NPC class", "NPC superclass", "NPC pathway", "ClassyFire subclass",
"ClassyFire class", "ClassyFire superclass"}
}
\value{
The dataList object with added sirius annotations.
}
\description{
Add qfeatures to dataList
}
