% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePeakAbundanceMatrixQF.R
\name{addSiriusAnnotations}
\alias{addSiriusAnnotations}
\title{Add Sirius Annotations}
\usage{
addSiriusAnnotations(
  qfeatures,
  siriusFile,
  rowData_col = "Alignment ID",
  sirius_col = "featureId"
)
}
\arguments{
\item{qfeatures}{dataset}

\item{siriusFile}{file path}

\item{rowData_col}{character Qfeatures column to match}

\item{sirius_col}{character Sirius column to match}
}
\value{
qfeatures object
}
\description{
Add Sirius Annotations
}
