% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeAlgorithms.R
\name{analyzeTree}
\alias{analyzeTree}
\title{Analyze tree}
\usage{
analyzeTree(
  dataList,
  cluster,
  filter,
  nodeIdx,
  minimumProportionOfLeafs = 0.75
)
}
\arguments{
\item{dataList}{list object}

\item{cluster}{cluster object}

\item{filter}{filter}

\item{nodeIdx}{numeric}
}
\value{
list object
}
\description{
Recursive function.
FIX creates many global variables
}
