% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotDendrogram}
\alias{calcPlotDendrogram}
\title{Prepare dendrogram plot}
\usage{
calcPlotDendrogram(
  dataList,
  filter,
  clusterDataList,
  annoPresentAnnotationsList,
  annoPresentColorsList,
  distanceMeasure,
  selectionFragmentTreeNodeSet = NULL,
  selectionAnalysisTreeNodeSet = NULL,
  selectionSearchTreeNodeSet = NULL,
  showClusterLabels,
  hcaPrecursorLabels,
  xInterval = NULL
)
}
\arguments{
\item{dataList}{}

\item{filter}{}

\item{clusterDataList}{}

\item{annoPresentAnnotationsList}{}

\item{annoPresentColorsList}{}

\item{distanceMeasure}{}

\item{selectionFragmentTreeNodeSet}{}

\item{selectionAnalysisTreeNodeSet}{}

\item{selectionSearchTreeNodeSet}{}

\item{showClusterLabels}{}

\item{hcaPrecursorLabels}{}

\item{xInterval}{}
}
\value{
?
}
\description{
Prepare dendrogram plot
}
