% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotDendrogram_plotly}
\alias{calcPlotDendrogram_plotly}
\title{Prepare Dendrogram for Plotly}
\usage{
calcPlotDendrogram_plotly(
  dataList,
  filterObj,
  clusterDataList,
  distanceMeasure,
  showClusterLabels,
  hcaPrecursorLabels,
  selectionFragmentTreeNodeSet = NULL,
  selectionAnalysisTreeNodeSet = NULL,
  selectionSearchTreeNodeSet = NULL,
  selectedSelection,
  heatmapContent,
  heatmapOrdering,
  heatmapProportion
)
}
\arguments{
\item{dataList}{}

\item{filterObj}{}

\item{clusterDataList}{}

\item{distanceMeasure}{}

\item{showClusterLabels}{}

\item{hcaPrecursorLabels}{}

\item{selectionFragmentTreeNodeSet}{}

\item{selectionAnalysisTreeNodeSet}{}

\item{selectionSearchTreeNodeSet}{}

\item{selectedSelection}{}

\item{heatmapContent}{}

\item{heatmapOrdering}{}

\item{heatmapProportion}{}
}
\value{
?
}
\description{
Prepare Dendrogram for Plotly
}
