% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotHeatmap}
\alias{calcPlotHeatmap}
\title{Plot Heatmap}
\usage{
calcPlotHeatmap(
  dataList,
  filterObj,
  clusterDataList,
  selectedTreeNodeSet,
  frameColor,
  heatmapContent,
  heatmapOrdering,
  xInterval = NULL
)
}
\arguments{
\item{dataList}{}

\item{filterObj}{}

\item{clusterDataList}{}

\item{selectedTreeNodeSet}{}

\item{frameColor}{}

\item{heatmapContent}{}

\item{heatmapOrdering}{}

\item{xInterval}{}
}
\value{
?
}
\description{
Plot Heatmap
}
