% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotMS2}
\alias{calcPlotMS2}
\title{MS2 Plot}
\usage{
calcPlotMS2(
  dataList,
  fragmentsX = NULL,
  fragmentsY = NULL,
  fragmentsColor = NULL,
  fragmentsDiscriminativity = NULL,
  fragmentsX_02 = NULL,
  fragmentsY_02 = NULL,
  fragmentsColor_02 = NULL,
  xInterval = NULL,
  selectedFragmentIndex = NULL,
  dendrogramFragmentStatistics = FALSE
)
}
\arguments{
\item{dataList}{dataList object}

\item{fragmentsX}{?}

\item{fragmentsY}{?}

\item{fragmentsColor}{?}

\item{fragmentsDiscriminativity}{?}

\item{fragmentsX_02}{?}

\item{fragmentsY_02}{?}

\item{fragmentsColor_02}{?}

\item{xInterval}{?}

\item{selectedFragmentIndex}{?}

\item{dendrogramFragmentStatistics}{?}
}
\value{
?
}
\description{
MS2 Plot
}
