% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotMS2Legend}
\alias{calcPlotMS2Legend}
\title{MS2 plot legend}
\usage{
calcPlotMS2Legend(
  dataList,
  minimumProportionOfLeafs = 0.75,
  minimumProportionToShowFragment = 0.5
)
}
\arguments{
\item{dataList}{list object}

\item{minimumProportionOfLeafs}{numeric}

\item{minimumProportionToShowFragment}{numeric}
}
\value{
?
}
\description{
MS2 plot legend
}
