% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotPCAloadings}
\alias{calcPlotPCAloadings}
\title{calcPlotPCAloadings}
\usage{
calcPlotPCAloadings(
  pcaObj,
  dataList,
  filter,
  pcaDimensionOne,
  pcaDimensionTwo,
  selectionFragmentPcaLoadingSet = NULL,
  selectionAnalysisPcaLoadingSet = NULL,
  selectionSearchPcaLoadingSet = NULL,
  xInterval = NULL,
  yInterval = NULL,
  loadingsLabels = "None",
  showLoadingsAbundance = FALSE,
  showLoadingsFeaturesAnnotated = TRUE,
  showLoadingsFeaturesUnannotated = TRUE,
  showLoadingsFeaturesSelected = TRUE,
  showLoadingsFeaturesUnselected = TRUE
)
}
\arguments{
\item{pcaObj}{}

\item{dataList}{}

\item{filter}{}

\item{pcaDimensionOne}{}

\item{pcaDimensionTwo}{}

\item{selectionFragmentPcaLoadingSet}{}

\item{selectionAnalysisPcaLoadingSet}{}

\item{selectionSearchPcaLoadingSet}{}

\item{xInterval}{}

\item{yInterval}{}

\item{loadingsLabels}{}

\item{showLoadingsAbundance}{}

\item{showLoadingsFeaturesAnnotated}{}

\item{showLoadingsFeaturesUnannotated}{}

\item{showLoadingsFeaturesSelected}{}

\item{showLoadingsFeaturesUnselected}{}
}
\value{
?
}
\description{
calcPlotPCAloadings
}
