% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotPCAscores}
\alias{calcPlotPCAscores}
\title{PCA plot}
\usage{
calcPlotPCAscores(
  pcaObj,
  dataList,
  filterObj,
  pcaDimensionOne,
  pcaDimensionTwo,
  showScoresLabels,
  xInterval = NULL,
  yInterval = NULL,
  downloadLayout = FALSE
)
}
\arguments{
\item{pcaObj}{}

\item{dataList}{}

\item{filterObj}{}

\item{pcaDimensionOne}{}

\item{pcaDimensionTwo}{}

\item{showScoresLabels}{}

\item{xInterval}{}

\item{yInterval}{}

\item{downloadLayout}{}
}
\value{
?
}
\description{
PCA plot
}
