% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{calculateCluster}
\alias{calculateCluster}
\title{Create cluster object}
\usage{
calculateCluster(
  dataList,
  filterObj,
  distanceMatrix,
  method,
  distanceMeasure,
  progress = FALSE
)
}
\arguments{
\item{dataList}{}

\item{filterObj}{}

\item{distanceMatrix}{}

\item{method}{}

\item{distanceMeasure}{}

\item{progress}{}
}
\value{
clusterDataList object
}
\description{
Create cluster object
}
