% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{castListEntries}
\alias{castListEntries}
\title{Cast logical's and numeric's in a list or data.frame}
\usage{
castListEntries(list)
}
\arguments{
\item{list}{}
}
\value{
list of the same lenght with logical's and numeric's casted
}
\description{
Tries to cast a list entry (or column in a data.frame) to logical's, 
if that does not create any missing values, it is assumed 
to be a logical will be replaced by `as.logical()` conversion.
Similarly for numeric entries (or columns). Everything else remains strings
}
