% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate cosine/modified cosine score}
\usage{
cosine_similarity(
  mz1,
  intensity1,
  precursor_mz1,
  mz2,
  intensity2,
  precursor_mz2,
  fragment_mz_tolerance = 0.01,
  allow_shift = TRUE,
  nl = FALSE,
  normalize = FALSE
)
}
\arguments{
\item{mz1}{numeric vector}

\item{intensity1}{numeric vector}

\item{precursor_mz1}{numeric}

\item{mz2}{numeric vector}

\item{intensity2}{numeric vector}

\item{precursor_mz2}{numeric}

\item{fragment_mz_tolerance}{numeric}

\item{allow_shift}{boolean}

\item{nl}{boolean}

\item{normalize}{boolean}
}
\value{
score
}
\description{
Calculate cosine/modified cosine score
}
