% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{data.numericmatrix}
\alias{data.numericmatrix}
\title{Convert data.frame columns to numeric}
\usage{
data.numericmatrix(x)
}
\arguments{
\item{x}{The data.frame to convert}
}
\value{
A matrix with all columns converted to numeric
}
\description{
The data.numericmatrix() function works similar to base::data.matrix()
before R-4.0.0 converting character columns to numeric without converting
to factor first, thus returning the actual numeric values.
}
\examples{
data.numericmatrix(data.frame(a = c("1", "2", "3"), 
                              b = c("4", "5", "6")))

}
