% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSP_refactor.R
\name{filter_msp_entries}
\alias{filter_msp_entries}
\title{Filter MSP entries}
\usage{
filter_msp_entries(
  msp_entries,
  minimumIntensityOfMaximalMS2peak,
  minimumProportionOfMS2peaks,
  neutralLossesPrecursorToFragments,
  neutralLossesFragmentsToFragments,
  mz_tol = 0.17,
  min_frag_nb = 1,
  min_intensity = 0
)
}
\arguments{
\item{msp_entries}{list of MSP entries created with \code{parseMSP_to_list()}}

\item{minimumIntensityOfMaximalMS2peak}{\code{numeric(1)} Spectra entries with a
maximum intensity below this value are removed.}

\item{minimumProportionOfMS2peaks}{\code{numeric(1)} Between 0 and 1. Within a
spectra, fragments with an intensity below this proportion of the maximum
intensity fragment are removed.}

\item{neutralLossesPrecursorToFragments}{boolean Should neutral losses to
precursor be calculated.}

\item{neutralLossesFragmentsToFragments}{boolean Should neutral losses for
each fragment pair be calculated.}

\item{mz_tol}{\code{numeric(1)} tolerance used for matching precursor fragment.}
}
\value{
list of spectra object, and stats
}
\description{
Should be used after \code{parseMSP_to_list()} to filter entries and MS2 spectra
based on a set of parameters.#'
}
