% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{filterData}
\alias{filterData}
\title{Filter a dataList}
\usage{
filterData(
  dataList,
  grouXXXps,
  sampleSet,
  filterBySamples,
  filter_average,
  filter_lfc,
  filterList_ms2_masses,
  filter_ms2_ppm,
  filter_ms1_masses,
  filter_ms1_ppm,
  includeIgnoredPrecursors,
  progress = FALSE
)
}
\arguments{
\item{dataList}{List object}

\item{grouXXXps}{sample class}

\item{sampleSet}{}

\item{filterBySamples}{}

\item{filter_average}{}

\item{filter_lfc}{}

\item{filterList_ms2_masses}{}

\item{filter_ms2_ppm}{}

\item{filter_ms1_masses}{}

\item{filter_ms1_ppm}{}

\item{includeIgnoredPrecursors}{}

\item{progress}{boolean}
}
\value{
A filtered dataList objetc
}
\description{
Filter a dataList
}
