% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{filterLFC}
\alias{filterLFC}
\title{Filter dataset based on LFC of two groups}
\usage{
filterLFC(dataList, filter_lfc, sampleClasses = dataList$sampleClasses)
}
\arguments{
\item{dataList}{list object}

\item{filter_lfc}{numeric e.g. 2 or -2}

\item{sampleClasses}{vector of names for group1 and group2}
}
\value{
boolean vector
}
\description{
Filter dataset based on LFC of two groups
}
