% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{filterThreshold}
\alias{filterThreshold}
\title{Filter dataset based on intensity mean values}
\usage{
filterThreshold(dataList, filter_average, grouXXXps = dataList$grouXXXps)
}
\arguments{
\item{dataList}{list object}

\item{filter_average}{numeric}

\item{grouXXXps}{vector of sample classes}
}
\value{
boolean vector
}
\description{
Filter dataset based on intensity mean values
}
