% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{getAvailableClassifiers}
\alias{getAvailableClassifiers}
\title{Classifiers}
\usage{
getAvailableClassifiers(resultFolderForClassifiers)
}
\arguments{
\item{resultFolderForClassifiers}{?}
}
\value{
?
}
\description{
Classifiers
}
