% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{getPrecursorColors}
\alias{getPrecursorColors}
\title{Create annotation colors}
\usage{
getPrecursorColors(dataList, precursorSet)
}
\arguments{
\item{dataList}{list object}

\item{precursorSet}{vector}
}
\value{
list with colors and annotations
}
\description{
Create annotation colors
}
