% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{impl_cosine_similarity}
\alias{impl_cosine_similarity}
\title{Implement Cosine Similarity}
\usage{
impl_cosine_similarity(
  dataList,
  filter,
  progress = FALSE,
  allow_shift = FALSE,
  nl = FALSE,
  num_filter = 5,
  rm_precursor = T
)
}
\arguments{
\item{dataList}{dataset}

\item{filter}{filter vector}

\item{progress}{boolena}

\item{allow_shift}{boolean, modified cosine when true}

\item{nl}{boolean, include neutral losses when true}

\item{num_filter}{integer, minimal number of fragments required to perform comparison}

\item{rm_precursor}{boolean, remove precursor ions from ms2 spectra}
}
\value{
distance matrix
}
\description{
Implement Cosine Similarity
}
