% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{make_filterObj}
\alias{make_filterObj}
\title{Create a filterObject for console use}
\usage{
make_filterObj(
  dataList,
  filter_average = NULL,
  filter_lfc = NULL,
  grouXXXps = NULL
)
}
\arguments{
\item{dataList}{list object}

\item{filter_average}{numeric}

\item{filter_lfc}{numeric}

\item{grouXXXps}{character vector}
}
\value{
filterObject
}
\description{
Mixes code from filterData() and doPerformFiltering_impl() to provide
a simple way to create a compatible filterObj in a script.
}
\details{
doPerformFiltering_impl() relies on a DataList in the global environment,
so it can't be used in a script.
}
