% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{metaboliteFamilyVersusClass}
\alias{metaboliteFamilyVersusClass}
\title{Title}
\usage{
metaboliteFamilyVersusClass(
  dataList,
  precursorSet,
  classToSpectra_class,
  properties_class,
  classifierClass,
  mappingSpectraToClassDf,
  addClassifierConsensusSpectrum
)
}
\arguments{
\item{dataList}{}

\item{precursorSet}{}

\item{classToSpectra_class}{}

\item{properties_class}{}

\item{classifierClass}{}

\item{mappingSpectraToClassDf}{}

\item{addClassifierConsensusSpectrum}{}
}
\value{
?
}
\description{
Title
}
