% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetFamily-package.R
\docType{package}
\name{MetFamily-package}
\alias{MetFamily}
\alias{MetFamily-package}
\title{MetFamily: MetFamily: Discovering Regulated Metabolite Families in Untargeted Metabolomics Studies}
\description{
Untargeted discovery of metabolite families offering a bird's eye view of metabolic regulation in comparative metabolomics. We implemented the presented methodology in the easy-to-use web application MetFamily to enable the analysis of comprehensive metabolomics studies for all researchers worldwide.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://ipb-halle.github.io/MetFamily/}
}

}
\keyword{internal}
