% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMGF.R
\name{mgf_dict}
\alias{mgf_dict}
\title{MGF header dictionnary}
\usage{
mgf_dict()
}
\value{
named vector with standardized names as value
}
\description{
Name matching for stadardized metadata.
Standard names based on \code{Spectra::spectraVariableMapping(MsBackendMsp::MsBackendMsp())}
}
