% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSP_refactor.R
\name{msp_dict}
\alias{msp_dict}
\title{MSP header dictionnary}
\usage{
msp_dict()
}
\value{
named vector with standardized names as value
}
\description{
Name matching for stadardized metadata.
Standard names based on \code{Spectra::spectraVariableMapping(MsBackendMsp::MsBackendMsp())}
}
