% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentMatrixFunctions.R
\name{mzClustGeneric}
\alias{mzClustGeneric}
\title{Alignment of high resolution mass spectra}
\usage{
mzClustGeneric(
  p,
  sampclass = NULL,
  mzppm = 20,
  mzabs = 0,
  minsamp = 1,
  minfrac = 0.5,
  progress = FALSE
)
}
\arguments{
\item{p}{?}

\item{sampclass}{sample class}

\item{mzppm}{numeric}

\item{mzabs}{numeric}

\item{minsamp}{numeric}

\item{minfrac}{numeric}

\item{progress}{boolean}
}
\value{
?
}
\description{
adapted from R package xcms: xcms_1.44.0, package path: R/mzClust.R
Reference:
Alignment of high resolution mass spectra: development of a heuristic approach for metabolomics
Metabolomics June 2006, Volume 2, Issue 2, pp 75-83
http://link.springer.com/article/10.1007\%2Fs11306-006-0021-7
}
