% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_project.R
\name{parameterSetDefault}
\alias{parameterSetDefault}
\title{Create a parameterSet object}
\usage{
parameterSetDefault()
}
\value{
a parameterSet list object of default values used by the app
}
\description{
To use in a script
}
