% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMGF.R
\name{parseMGF}
\alias{parseMGF}
\title{Read MGF files}
\usage{
parseMGF(
  fileSpectra,
  minimumIntensityOfMaximalMS2peak = 2000,
  minimumProportionOfMS2peaks = 0.05,
  neutralLossesPrecursorToFragments = T,
  neutralLossesFragmentsToFragments = F,
  mz_tol = 0.17,
  min_frag_nb = 2,
  min_intensity = 500,
  progress = FALSE
)
}
\arguments{
\item{fileSpectra}{\code{character(1)} file path}

\item{minimumIntensityOfMaximalMS2peak}{\code{numeric(1)} Spectra entries with a
maximum intensity below this value are removed.}

\item{minimumProportionOfMS2peaks}{\code{numeric(1)} Between 0 and 1. Within a
spectra, fragments with an intensity below this proportion of the maximum
intensity fragment are removed.}

\item{neutralLossesPrecursorToFragments}{boolean Should neutral losses to
precursor be calculated.}

\item{neutralLossesFragmentsToFragments}{boolean Should neutral losses for
each fragment pair be calculated.}

\item{mz_tol}{\code{numeric(1)} tolerance used for matching precursor fragment.}

\item{min_frag_nb}{\code{numeric(1)} minimum number of fragments needed to include spectrum}

\item{min_intensity}{\code{numeric(1)} minimum intensity to include fragment}

\item{progress}{boolean, used for interactive app.}
}
\value{
A list including a spectraList element and additional stats.
}
\description{
Reader for Mascot Generic Format MS2 files. Works similar to \code{parseMSP_rewrite()}.
}
