% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMGF.R
\name{parseMGF_to_list}
\alias{parseMGF_to_list}
\title{Process MGF from file to list of entries}
\usage{
parseMGF_to_list(fileSpectra)
}
\arguments{
\item{fileSpectra}{\code{character(1)} file path}
}
\value{
list of entries
}
\description{
Create list object from MGF entry, keeping all headers and values
Rename headers based on standards
}
