% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentMatrixFunctions.R
\name{parseMSP_attributes}
\alias{parseMSP_attributes}
\title{parseMSP_attributes}
\usage{
parseMSP_attributes(
  fileSpectra,
  progress = FALSE,
  flexiblePeakList = FALSE,
  multiplePeaksPerLine = FALSE,
  includeIDasRecordSeparator = TRUE,
  includeNAMEasRecordSeparator = TRUE,
  includeTITLEasRecordSeparator = TRUE,
  returnEmptySpectra = FALSE
)
}
\arguments{
\item{fileSpectra}{}

\item{progress}{}

\item{flexiblePeakList}{}

\item{multiplePeaksPerLine}{}

\item{includeIDasRecordSeparator}{}

\item{includeNAMEasRecordSeparator}{}

\item{includeTITLEasRecordSeparator}{}

\item{returnEmptySpectra}{}
}
\value{
?
}
\description{
parseMSP_attributes
}
