% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentMatrixFunctions.R
\name{parseMSP_chunk}
\alias{parseMSP_chunk}
\title{parse MSP file to list}
\usage{
parseMSP_chunk(
  fileLines,
  minimumIntensityOfMaximalMS2peak,
  minimumProportionOfMS2peaks,
  neutralLossesPrecursorToFragments,
  neutralLossesFragmentsToFragments,
  offset = 0,
  progress = FALSE
)
}
\arguments{
\item{fileLines}{}

\item{minimumIntensityOfMaximalMS2peak}{numeric}

\item{minimumProportionOfMS2peaks}{numeric}

\item{neutralLossesPrecursorToFragments}{boolean}

\item{neutralLossesFragmentsToFragments}{boolean}

\item{offset}{}

\item{progress}{boolean}
}
\value{
list
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Superseded with parseMSP rewrite in 2025-08
}
