% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSP_refactor.R
\name{parseMSP_to_list}
\alias{parseMSP_to_list}
\title{Process MSP from file to list of entries}
\usage{
parseMSP_to_list(fileSpectra)
}
\arguments{
\item{fileSpectra}{\code{character(1)} file path}
}
\value{
list of entries
}
\description{
Create list object from MSP entry, keeping all headers and values
Rename headers based on standards
}
