% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentMatrixFunctions.R
\name{parsePeakAbundanceMatrix}
\alias{parsePeakAbundanceMatrix}
\title{Read and preprocess MS1 peak abundance matrix}
\usage{
parsePeakAbundanceMatrix(
  filePeakMatrix,
  doPrecursorDeisotoping,
  mzDeviationInPPM_precursorDeisotoping,
  mzDeviationAbsolute_precursorDeisotoping,
  maximumRtDifference,
  progress
)
}
\arguments{
\item{filePeakMatrix}{}

\item{doPrecursorDeisotoping}{}

\item{mzDeviationInPPM_precursorDeisotoping}{}

\item{mzDeviationAbsolute_precursorDeisotoping}{}

\item{maximumRtDifference}{}

\item{progress}{whether or not to use and update a progress bar when running under shiny}
}
\value{
Heck, we need to document that.
}
\description{
Read and preprocess MS1 peak abundance matrix
}
\examples{
  filePeakMatrix <- system.file("data/showcase/Metabolite_profile_showcase.txt", 
                                package = "MetFamily")
  data <- parsePeakAbundanceMatrix(filePeakMatrix, 
                                   doPrecursorDeisotoping=TRUE,
                                   mzDeviationInPPM_precursorDeisotoping=10, 
                                   mzDeviationAbsolute_precursorDeisotoping=0.01,
                                   maximumRtDifference=0.05,
                                   progress=FALSE)
}
