% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePeakAbundanceMatrixQF.R
\name{parsePeakAbundanceMatrixQF}
\alias{parsePeakAbundanceMatrixQF}
\title{Title}
\usage{
parsePeakAbundanceMatrixQF(
  qfeatures,
  doPrecursorDeisotoping,
  mzDeviationInPPM_precursorDeisotoping,
  mzDeviationAbsolute_precursorDeisotoping,
  maximumRtDifference,
  progress = FALSE
)
}
\arguments{
\item{qfeatures}{Object of type QFeatures}

\item{doPrecursorDeisotoping}{boolean}

\item{mzDeviationInPPM_precursorDeisotoping}{numeric}

\item{mzDeviationAbsolute_precursorDeisotoping}{numeric}

\item{maximumRtDifference}{numeric}

\item{progress}{boolean whether or not to show a progress bar}
}
\value{
?
}
\description{
Title
}
