% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{performPca}
\alias{performPca}
\title{Run PCA}
\usage{
performPca(dataList, dataFrame2, ms1AnalysisMethod)
}
\arguments{
\item{dataList}{List object}

\item{dataFrame2}{}

\item{ms1AnalysisMethod}{}
}
\value{
A PCA list object
}
\description{
Run PCA
}
