% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSP_refactor.R
\name{process_msp_entry}
\alias{process_msp_entry}
\title{Filter single MSP entry}
\usage{
process_msp_entry(
  msp_entry,
  minimumProportionOfMS2peaks,
  neutralLossesPrecursorToFragments,
  neutralLossesFragmentsToFragments,
  min_intensity = 500
)
}
\arguments{
\item{msp_entry}{MSP entry object}

\item{minimumProportionOfMS2peaks}{\code{numeric(1)} Between 0 and 1. Within a
spectra, fragments with an intensity below this proportion of the maximum
intensity fragment are removed.}

\item{neutralLossesPrecursorToFragments}{boolean Should neutral losses to
precursor be calculated.}

\item{neutralLossesFragmentsToFragments}{boolean Should neutral losses for
each fragment pair be calculated.}

\item{min_intensity}{\code{numeric(1)} minimum intensity to include fragment}
}
\value{
filtered spectra
}
\description{
nb of peaks and peaks with NL are updated in header
peaks are filtered, and neutral losses added if requested
stats section added
returns backwards-compatible spectraItem as used by MetFamily
}
