% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{projectFromFiles}
\alias{projectFromFiles}
\title{Create a project dataList}
\usage{
projectFromFiles(
  ms1_path,
  ms2_path,
  annot_path = NULL,
  siriusFileColumnName = c("NPC class"),
  parameterSet = NULL
)
}
\arguments{
\item{ms1_path}{file for MS1 intensity .txt file}

\item{ms2_path}{file for MS2 .msp file}

\item{annot_path}{file for annotations (.csv or .tsv)}

\item{parameterSet}{list of parameters to use}
}
\value{
dataList
}
\description{
Wrapper for a clunky workflow
}
