% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMetaboScape.R
\name{readMetaboscape}
\alias{readMetaboscape}
\title{Read Metaboscape Output File into a QFeatures Object}
\usage{
readMetaboscape(file, version)
}
\arguments{
\item{file}{A character string specifying the path to the Metaboscape output file (Excel format).}

\item{version}{A character string specifying the version of Metaboscape used to generate the file.
This parameter is currently not used.}
}
\value{
A QFeatures object containing:
\itemize{
\item An assay named "exampleAssay" with the metabolite counts.
\item Row data (feature metadata) extracted from the input file.
\item Column data (sample metadata) extracted from the sample names, including injection order and sample name.
}
}
\description{
This function reads a metabolite profile output file (.xlsx) from Metaboscape and
converts it into a QFeatures object.
}
\examples{
\dontrun{
# Assuming you have a Metaboscape output file named "data.xlsx":
qf <- readMetaboscape("data.xlsx") #TODO: System file 


# Examine the structure of the resulting QFeatures object
qf

# Access the assay data
assay(qf[["exampleAssay"]])

# Access the row data (feature metadata)
rowData(qf[["exampleAssay"]])

# Access the column data (sample metadata)
colData(qf)
}


}
\references{
#TODO: Bruker metaboscape site
#TODO: Ordering ?
}
\seealso{
\code{\link[QFeatures]{QFeatures}} for more information on the QFeatures class.
\code{\link[SummarizedExperiment]{SummarizedExperiment}} for details on the underlying data structure.
}
