% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMSDial.R
\name{readMSDial}
\alias{readMSDial}
\title{Read MS-DIAL Output File into a QFeatures Object}
\usage{
readMSDial(file, version)
}
\arguments{
\item{file}{A string with the path to the MS-DIAL output file.}

\item{version}{A character string specifying the version of MS-DIAL used to
generate the file. This parameter is currently not used.}
}
\value{
A QFeatures object containing:
\itemize{
\item An assay named "exampleAssay" with the metabolite counts.
\item Row data (feature metadata) extracted from the input file.
\item Column data (sample metadata) extracted from the input file.
}
}
\description{
This function reads the output file from MS-DIAL and converts it into a
QFeatures object.
}
\examples{
\dontrun{
# Assuming you have an MS-DIAL output file named
 "Metabolite_profile_showcase.txt" in a "data" directory:
qf <- readMSDial("data/Metabolite_profile_showcase.txt")

# Examine the structure of the resulting QFeatures object
qf

# Access the assay data
assay(qf[["exampleAssay"]])

# Access the row data (feature metadata)
rowData(qf[["exampleAssay"]])

# Access the column data (sample metadata)
colData(qf)
}

}
\seealso{
\code{\link[QFeatures]{QFeatures}} for more information on the
QFeatures class. \code{\link[SummarizedExperiment]{SummarizedExperiment}} for
details on the underlying data structure.
}
