% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMZmine.R
\name{readMZmine}
\alias{readMZmine}
\title{Read mzmine Output File into a QFeatures Object}
\usage{
readMZmine(file, version)
}
\arguments{
\item{file}{A character string specifying the path to the Metaboscape output file (csv format).}

\item{version}{A character string specifying the version of Metaboscape used to generate the file.
This parameter is currently not used.}
}
\value{
A QFeatures object containing:
\itemize{
\item An assay named "exampleAssay" with the metabolite counts.
\item Row data (feature metadata) extracted from the input file.
\item Column data (sample metadata) extracted from the sample names, including injection order and sample name.
}
}
\description{
Read a metabolite profile output file (.csv) from mzmine and
converts it into a QFeatures object.
}
\details{
At the moment, sample classes are not considered.
}
