% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_project.R
\name{readProjectFile}
\alias{readProjectFile}
\title{Read MetFamily Project data saved by the export function}
\usage{
readProjectFile(file, progress = FALSE)
}
\arguments{
\item{file}{Path to file to read}

\item{progress}{Whether to update a shiny Progress bar}
}
\value{
A big dataList.
}
\description{
Supports reading from plain and gzip'ed files
}
\seealso{
\link{readProjectData}
}
