% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{sparseMatrixToString}
\alias{sparseMatrixToString}
\title{Flatten listObject to string for export}
\usage{
sparseMatrixToString(matrixRows, matrixCols, matrixVals, parameterSet)
}
\arguments{
\item{matrixRows}{}

\item{matrixCols}{}

\item{matrixVals}{}

\item{parameterSet}{}
}
\value{
big string object
}
\description{
Also used before running readProjectData
}
