% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartApp.R
\name{startMetFamily}
\alias{startMetFamily}
\title{Start the web app in browser}
\usage{
startMetFamily()
}
\description{
Start the web app in browser
}
\examples{
startMetFamily()
}
\author{
Hendrik Treutler
}
