% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{swap_names_and_values}
\alias{swap_names_and_values}
\title{Swap a vector's names and values}
\usage{
swap_names_and_values(x)
}
\arguments{
\item{x}{vector}
}
\value{
string
}
\description{
Similar to \code{searchable::invert()}
}
