% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_project.R
\name{writeProjectFile}
\alias{writeProjectFile}
\title{Export MetFamily project}
\usage{
writeProjectFile(dataList, precursorSet = NULL, file)
}
\arguments{
\item{dataList}{main dataset}

\item{precursorSet}{entries to include, defaults to all}

\item{file}{filename}
}
\value{
filename
}
\description{
Creates a \code{.gz} file to store the MetFamily project.
}
\details{
Implements the logic of prepareAllPrecursors in a generalized manner.
Replaces \code{createExportMatrix()} as a package function.
}
