% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{add_qfeatures}
\alias{add_qfeatures}
\title{Add qfeatures to dataList}
\usage{
add_qfeatures(
  dataList,
  qfeatures,
  fileAnnotation = NULL,
  siriusCategory = "NPC class"
)
}
\arguments{
\item{dataList}{Output from readProjectData.}

\item{qfeatures}{qfeature object, can be taken from resultObj$qfeatures.}

\item{fileAnnotation}{character Path for sirius annotation file.}
}
\value{
The dataList object with added sirius annotations.
}
\description{
Add qfeatures to dataList
}
