% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePeakAbundanceMatrixQF.R
\name{addSiriusAnnotations}
\alias{addSiriusAnnotations}
\title{Add Sirius Annotations}
\usage{
addSiriusAnnotations(
  qfeatures,
  siriusFile,
  rowData_col = "Alignment ID",
  sirius_col = "featureId"
)
}
\arguments{
\item{qfeatures}{}

\item{siriusFile}{}

\item{rowData_col}{}

\item{sirius_col}{}
}
\value{
?
}
\description{
Add Sirius Annotations
}
