% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePeakAbundanceMatrixQF.R
\name{addSiriusAnnotations}
\alias{addSiriusAnnotations}
\title{Add Sirius Annotations}
\usage{
addSiriusAnnotations(
  qfeatures,
  siriusFile,
  rowData_col = "Alignment ID",
  sirius_col = "featureId",
  siriusFileColumnName = "NPC class"
)
}
\arguments{
\item{qfeatures}{dataset}

\item{siriusFile}{file path}

\item{rowData_col}{character Qfeatures column to match}

\item{sirius_col}{character Sirius column to match}

\item{siriusFileColumnName}{One of "NPC class", "NPC superclass", "NPC pathway", "ClassyFire subclass",
"ClassyFire class", "ClassyFire superclass"}
}
\value{
qfeatures object
}
\description{
Missing values in the Sirius file create empty character values "" in the annotation,
while missing features have NA values. All are transformed to "" later in \code{add_qfeatures}.
}
