% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotPCAloadings}
\alias{calcPlotPCAloadings}
\title{PCA plot of loadings}
\usage{
calcPlotPCAloadings(
  pcaObj,
  dataList,
  filterVec = NULL,
  pcaDimensionOne = 1,
  pcaDimensionTwo = 1,
  selectionFragmentPcaLoadingSet = NULL,
  selectionAnalysisPcaLoadingSet = NULL,
  selectionSearchPcaLoadingSet = NULL,
  xInterval = NULL,
  yInterval = NULL,
  loadingsLabels = "None",
  showLoadingsAbundance = FALSE,
  showLoadingsFeaturesAnnotated = TRUE,
  showLoadingsFeaturesUnannotated = TRUE,
  showLoadingsFeaturesSelected = TRUE,
  showLoadingsFeaturesUnselected = TRUE
)
}
\arguments{
\item{pcaObj}{output from function calculatePCA}

\item{dataList}{dataList object}

\item{filterVec}{deprecated, use filterObj element in calculatePCA() instead}

\item{pcaDimensionOne}{numeric PCA dimension for x axis}

\item{pcaDimensionTwo}{numeric PCA dimension for y axis}

\item{selectionFragmentPcaLoadingSet}{numeric vector of indices}

\item{selectionAnalysisPcaLoadingSet}{numeric vector of indices}

\item{selectionSearchPcaLoadingSet}{numeric vector of indices}

\item{xInterval}{numeric Plot parameter}

\item{yInterval}{numeric Plot parameter}

\item{loadingsLabels}{boolean One of "None", "m/z / RT", "Metabolite name", or "Metabolite family"}

\item{showLoadingsAbundance}{logical}

\item{showLoadingsFeaturesAnnotated}{logical}

\item{showLoadingsFeaturesUnannotated}{logical}

\item{showLoadingsFeaturesSelected}{logical}

\item{showLoadingsFeaturesUnselected}{logical}
}
\value{
a resultList object with annotation colors
}
\description{
Show feature loadings on a PCA plot.
}
