% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{calcPlotPCAscores}
\alias{calcPlotPCAscores}
\title{PCA plot of samples}
\usage{
calcPlotPCAscores(
  pcaObj,
  dataList,
  filterObj,
  pcaDimensionOne = 1,
  pcaDimensionTwo = 2,
  showScoresLabels,
  xInterval = NULL,
  yInterval = NULL,
  downloadLayout = FALSE
)
}
\arguments{
\item{pcaObj}{list object}

\item{dataList}{list object}

\item{filterObj}{filter obj}

\item{pcaDimensionOne}{first PCA axis}

\item{pcaDimensionTwo}{second PCA axis}

\item{showScoresLabels}{}

\item{xInterval}{numeric}

\item{yInterval}{numeric}

\item{downloadLayout}{boolean, set to TRUE for a different figure better suited for printing}
}
\value{
makes a plot, returns NULL
}
\description{
PCA plot of samples
}
