% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{calculatePCA}
\alias{calculatePCA}
\title{Calculate PCA}
\usage{
calculatePCA(dataList, filterObj, ms1AnalysisMethod, scaling, logTransform)
}
\arguments{
\item{dataList}{listObject}

\item{filterObj}{filterObject}

\item{ms1AnalysisMethod}{}

\item{scaling}{}

\item{logTransform}{}
}
\value{
PCA object
}
\description{
Calculate PCA
}
