% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{filterData}
\alias{filterData}
\title{Filter a dataList}
\usage{
filterData(
  dataList,
  grouXXXps,
  sampleSet = NULL,
  filterBySamples = FALSE,
  filter_average = NULL,
  filter_lfc = NULL,
  filterList_ms2_masses = NULL,
  filter_ms2_ppm = NULL,
  filter_ms1_masses = NULL,
  filter_ms1_ppm = NULL,
  includeIgnoredPrecursors = FALSE,
  progress = FALSE
)
}
\arguments{
\item{dataList}{List object}

\item{grouXXXps}{sample class}

\item{sampleSet}{?}

\item{filterBySamples}{boolean}

\item{filter_average}{numeric Features that have a mean intensity below that threshold are filtered out.}

\item{filter_lfc}{numeric Log-fold-change threshold between two treatments. Only implement when exactly two groups are present.}

\item{filterList_ms2_masses}{?}

\item{filter_ms2_ppm}{?}

\item{filter_ms1_masses}{?}

\item{filter_ms1_ppm}{?}

\item{includeIgnoredPrecursors}{boolean}

\item{progress}{boolean}
}
\value{
A filtered dataList object
}
\description{
Filter a dataList
}
