% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{impl_cosine_similarity}
\alias{impl_cosine_similarity}
\title{Implement Cosine Similarity}
\usage{
impl_cosine_similarity(
  dataList,
  filter,
  progress = FALSE,
  allow_shift = FALSE,
  nl = FALSE
)
}
\arguments{
\item{dataList}{dataset}

\item{filter}{filter vector}

\item{progress}{boolena}

\item{allow_shift}{boolean}

\item{nl}{boolean}
}
\value{
distance matrix
}
\description{
Implement Cosine Similarity
}
