% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentMatrixFunctions.R
\name{parseMSP}
\alias{parseMSP}
\title{parse MS/MS spectra}
\usage{
parseMSP(
  fileSpectra,
  minimumIntensityOfMaximalMS2peak,
  minimumProportionOfMS2peaks,
  neutralLossesPrecursorToFragments,
  neutralLossesFragmentsToFragments,
  progress = FALSE
)
}
\arguments{
\item{fileSpectra}{file}

\item{minimumIntensityOfMaximalMS2peak}{numeric}

\item{minimumProportionOfMS2peaks}{numeric}

\item{neutralLossesPrecursorToFragments}{boolean}

\item{neutralLossesFragmentsToFragments}{boolean}

\item{progress}{boolean}
}
\value{
list
}
\description{
parse MS/MS spectra
}
