% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentMatrixFunctions.R
\name{parseMSP}
\alias{parseMSP}
\title{Parse MS/MS spectra}
\usage{
parseMSP(
  fileSpectra,
  minimumIntensityOfMaximalMS2peak,
  minimumProportionOfMS2peaks,
  neutralLossesPrecursorToFragments,
  neutralLossesFragmentsToFragments,
  progress = FALSE,
  ...
)
}
\arguments{
\item{fileSpectra}{filepath}

\item{minimumIntensityOfMaximalMS2peak}{numeric}

\item{minimumProportionOfMS2peaks}{numeric}

\item{neutralLossesPrecursorToFragments}{boolean}

\item{neutralLossesFragmentsToFragments}{boolean}

\item{progress}{boolean}

\item{...}{ignored. Used for forward compatibility.}
}
\value{
list object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Superseded by parseMSP rewrite 2025-08
}
\details{
Read MS2 spectra from MSP file
}
