% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{readProjectData}
\alias{readProjectData}
\title{Read MetFamily Project data saved by the export function}
\usage{
readProjectData(fileLines, progress = FALSE)
}
\arguments{
\item{fileLines}{Character vector with content of a project file}

\item{progress}{Whether to update a shiny Progress bar}
}
\value{
A big dataList.
}
\description{
Read MetFamily Project data saved by the export function
}
\seealso{
\link{processMS1data}
}
