% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetFamily.R
\name{runMetFamily}
\alias{runMetFamily}
\title{Start the web app in browser}
\usage{
runMetFamily()
}
\description{
Run the web app in browser
}
\examples{
\dontrun{
runMetFamily()
}
}
\author{
Hendrik Treutler, Steffen Neumann
}
